chore: rename capacitor folder to native
