+++ apps/web/src/app/_provider/index.ts
@@ -1,4 +1,4 @@
 export { AuthRefreshProvider } from './auth-refresh';
 export { GlobalRoutingProvider } from './global-routing';
-export { ForceUpdateProvider } from './capacitor';
+export { ForceUpdateProvider } from './native';
 export { OnboardingGuard } from './onboarding-guard';
