+++ apps/web/src/features/auth/model/mutations/useRequestNativeSocialTokenMutation.ts
@@ -8,7 +8,7 @@ import {
   requestNativeSocialLogin,
   type NativeSocialLoginToken,
   type SocialProvider,
-} from '@/shared/lib/capacitor';
+} from '@/shared/lib/native';
 
 type NativeSocialBridgeVariables = Pick<
   NativeSocialLoginRequestDto,
