+++ apps/web/src/shared/lib/index.ts
@@ -1,5 +1,5 @@
 export * from './cloudfront';
-export * from './capacitor';
+export * from './native';
 export * from './daum-postcode';
 export * from './image-save';
 export * from './format';
