+++ apps/web/src/shared/lib/native/app-version/compareVersion.ts
