+++ apps/web/src/shared/lib/native/app-version/fetchMinVersion.ts
