+++ apps/web/src/shared/lib/native/app-version/forceUpdate.ts
