+++ apps/web/src/shared/lib/native/app-version/getAppVersion.ts
