+++ apps/web/src/shared/lib/native/app-version/index.ts
