+++ apps/web/src/shared/lib/native/app-version/openAppStore.ts
