+++ apps/web/tsconfig.json
@@ -11,7 +11,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "src/shared/lib/capacitor"
+    "src/shared/lib/native"
   ],
   "exclude": ["node_modules"]
 }
